fixes for fps

+++ LePong/constants.h
@@ -1,9 +1,9 @@
 #include <raylib.h>
 
-extern const int BALL_RADIUS;
+extern const float BALL_RADIUS;
 extern const Vector2 BALL_VELOCITY;
 extern const float BALL_ACCELERATION;
 extern const float MAX_BALL_ACCELERATION;
 extern const Vector2 PADDLE_SIZE;
 extern const float PADDLE_MARGIN;
-extern const int PADDLE_VELOCITY; 
+extern const float PADDLE_VELOCITY;
